backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
werkzeug==3.0.3
pyjwt==2.9.0
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
